% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{centroid}
\alias{centroid}
\title{Get the coordinates of the center of a box.}
\usage{
centroid(b)
}
\arguments{
\item{b}{A numeric vector representing a box like \code{c(x1, y1, x2, y2)}}
}
\description{
Get the coordinates of the center of a box.
}

