% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{euclid}
\alias{euclid}
\title{Euclidean distance between two numeric vectors.}
\usage{
euclid(p1, p2)
}
\arguments{
\item{p1}{A numeric vector.}

\item{p2}{A numeric vector.}
}
\description{
Euclidean distance between two numeric vectors.
}

