% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-image-repel.R
\name{geom_image_repel}
\alias{geom_image_repel}
\title{Plot Repeled Images.}
\usage{
geom_image_repel(mapping = NULL, data = NULL, stat = "identity",
  parse = FALSE, ..., box.padding = unit(0.25, "lines"),
  point.padding = unit(1e-06, "lines"), width = 0.1, aspectratio = 1,
  segment.size = 0.5, segment.alpha = 0.5, arrow = NULL, force = 1,
  max.iter = 2000, nudge_x = 0, nudge_y = 0, na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2]{aes}} or
\code{\link[ggplot2]{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), is combined with the default mapping at the top level of the
plot.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. There are
  three types of arguments you can use here:


  \itemize{
    \item Aesthetics: to set an aesthetic to a fixed value, like
       \code{color = "red"} or \code{size = 3}.
    \item Other arguments to the layer, for example you override the
      default \code{stat} associated with the layer.
    \item Other arguments passed on to the stat.
  }}

\item{width}{the width of the image files to plot using "npc" units where 1=100%
Defaults to \code{unit(0.1, "npc")}.}

\item{segment.size}{Width of line segment connecting the data point to
the text label, in mm.}

\item{segment.alpha}{Transparency of the segment, in \code{[0,1]}. Makes segments half transparent by default.}

\item{arrow}{specification for arrow heads, as created by \code{\link[grid]{arrow}}}

\item{force}{Force of repulsion between overlapping text labels. Defaults
to 1.}

\item{max.iter}{Maximum number of iterations to try to resolve overlaps.
Defaults to 2000.}

\item{nudge_x, nudge_y}{Horizontal and vertical adjustments to nudge the
starting position of each text label.}

\item{image}{The name of the column holding filepaths to you PNG images.}

\item{aspectratio.}{Along with width it is used to determine the width of your imgae. Defaults to
\code{1}.}
}
\description{
\code{geom_image_repel} adds images directly to the plot after moving them
from the starting position to avoid overlap. This code uses all of the options
available to \code{\link{geom_label_repel}} and \code{\link{geom_text_repel}} in
addtion to a few image-specific parameters. Currently this only supports the plotting
of PNG images.
}
\section{\code{geom_image_repel}}{

}
\examples{

#add a column with text files corresponding to an image location
rpng <- system.file("img", "Rlogo.png", package="png")
mtcars$images <- rpng

p <- ggplot(mtcars, aes(mpg,wt)) + geom_point()

# Add images to the plote for each point
p + geom_image_repel(aes(image=images))

# control the size and shape of images using the
# width and aspect ratio
p + geom_image_repel(aes(image=images), width=0.03)
p + geom_image_repel(aes(image=images), aspectratio=0.2)

# Label subsets of the data
p + geom_image_repel(data= mtcars[mtcars$wt>10,], aes(image=images))

}

