% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-text-wordcloud.R
\name{geom_text_wordcloud}
\alias{geom_text_wordcloud}
\title{Wordcloud.}
\usage{
geom_text_wordcloud(mapping = NULL, data = NULL, stat = "identity",
  parse = FALSE, ..., box.padding = 0.025, point.padding = NA,
  segment.colour = NA, segment.color = NA, segment.size = 0.5,
  segment.alpha = NULL, min.segment.length = 0.5, arrow = NULL,
  rstep = 0.1, tstep = 0.1, max.iter = 2000, nudge_x = 0, nudge_y = 0,
  xlim = c(NA, NA), ylim = c(NA, NA), na.rm = FALSE, show.legend = NA,
  seed = NA, inherit.aes = TRUE)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2]{aes}} or
\code{\link[ggplot2]{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), is combined with the default mapping at the top level of the
plot. You only need to supply \code{mapping} if there isn't a mapping
defined for the plot.}

\item{data}{A data frame. If specified, overrides the default data frame
defined at the top level of the plot.}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{parse}{If TRUE, the labels will be parsed into expressions and
displayed as described in ?plotmath}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. There are
  three types of arguments you can use here:

  \itemize{
    \item Aesthetics: to set an aesthetic to a fixed value, like
       \code{colour = "red"} or \code{size = 3}.
    \item Other arguments to the layer, for example you override the
      default \code{stat} associated with the layer.
    \item Other arguments passed on to the stat.
  }}

\item{box.padding}{Amount of padding around bounding box, as unit or number.
Defaults to 0.25. (Default unit is lines, but other units can be specified
by passing \code{unit(x, "units")}).}

\item{point.padding}{Amount of padding around labeled point, as unit or
number. Defaults to 0. (Default unit is lines, but other units can be
specified by passing \code{unit(x, "units")}).}

\item{segment.colour, segment.color}{Colour of the line segment. Defaults to the same colour
as the text. In the unlikely event you specify both US and UK spellings of colour, the
US spelling will take precedence. Default is NA so that no segment are drawn}

\item{segment.size}{Width of line segment connecting the data point to
the text label, in mm.}

\item{segment.alpha}{Transparency of the line segment. Defaults to the same
transparency as the text.}

\item{min.segment.length}{Skip drawing segments shorter than this, as unit or
number. Defaults to 0.5. (Default unit is lines, but other units can be
specified by passing \code{unit(x, "units")}).}

\item{arrow}{specification for arrow heads, as created by \code{\link[grid]{arrow}}}

\item{rstep}{wordclould spiral radius increment after one full rotation. Default to .1.}

\item{tstep}{wordclould spiral angle increment at each step. Default to .1.}

\item{max.iter}{Maximum number of iterations to try to resolve overlaps.
Defaults to 2000.}

\item{nudge_x, nudge_y}{Horizontal and vertical adjustments to nudge the
starting position of each text label.}

\item{xlim, ylim}{Limits for the x and y axes. Text labels will be constrained
to these limits. By default, text labels are constrained to the entire plot
area.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{seed}{Random seed passed to \code{\link[base]{set.seed}}. Defaults to
\code{NA}, which means that \code{set.seed} will not be called.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2]{borders}}.}
}
\description{
\code{geom_text_wordcloud} adds text directly to the plot.
}
\details{
This geom is based on \code{\link[ggplot2]{geom_text}}. See the
documentation for those functions for more details. Differences
 from those functions are noted here.

Text labels have height and width, but they are physical units, not data
units. The amount of space they occupy on that plot is not constant in data
units: when you resize a plot, labels stay the same size, but the size of
the axes changes. The text labels are repositioned after resizing a plot.
}
\section{Alignment}{

The repulsive geoms reposition text labels to avoid overlap, so the
following parameters are \strong{not supported}:

\itemize{
  \item \code{position}
  \item \code{check_overlap}
}

The parameters \code{hjust} and \code{vjust} are supported, but they only
control the intial positioning and repulsive forces may disrupt alignment.
Alignment will be preserved if the force is limited to not act in the
direction of alignment using the \code{direction} parameter.
}

\examples{
p <- ggplot(mtcars,
            aes(label = rownames(mtcars), size = mpg,
                colour = factor(cyl)))

p + geom_text_wordcloud()

p + geom_text_wordcloud() +
  scale_size(limits = c(0, NA))

p + geom_text_wordcloud(aes(angle = 90 * (runif(nrow(mtcars))>.9)))

p + geom_text_wordcloud(aes(x = cyl)) +
  scale_x_continuous(limits = c(3, 9))
}
