% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-label-repel.R, R/geom-text-repel.R,
%   R/position-nudge2.R
\docType{data}
\name{GeomLabelRepel}
\alias{GeomLabelRepel}
\alias{GeomTextRepel}
\alias{PositionNudge2}
\title{GeomLabelRepel}
\description{
GeomLabelRepel

GeomTextRepel
}
\keyword{datasets}
