% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-label-repel.R, R/geom-marquee-repel.R,
%   R/geom-text-repel.R, R/ggrepel-package.R, R/position-nudge-repel.R
\docType{data}
\name{GeomLabelRepel}
\alias{GeomLabelRepel}
\alias{GeomMarqueeRepel}
\alias{GeomTextRepel}
\alias{ggrepel}
\alias{PositionNudgeRepel}
\title{GeomLabelRepel}
\description{
This package contains extra geoms for \pkg{ggplot2}.
}
\details{
Please see the help pages listed below:

\itemize{
  \item \code{\link{geom_text_repel}}
  \item \code{\link{geom_label_repel}}
}

Also see the vignette for more usage examples:

\code{browseVignettes("ggrepel")}

Please report issues and suggest improvements at Github:

\url{https://github.com/slowkow/ggrepel}
}
\seealso{
\link[ggplot2]{GeomLabel} from the ggplot2 package.

\link[ggplot2]{GeomText} from the ggplot2 package.

Useful links:
\itemize{
  \item \url{https://ggrepel.slowkow.com/}
  \item \url{https://github.com/slowkow/ggrepel}
  \item Report bugs at \url{https://github.com/slowkow/ggrepel/issues}
}

}
\author{
\strong{Maintainer}: Kamil Slowikowski \email{kslowikowski@gmail.com} (\href{https://orcid.org/0000-0002-2843-6370}{ORCID})

Other contributors:
\itemize{
  \item Teun van den Brand (\href{https://orcid.org/0000-0002-9335-7468}{ORCID}) [contributor]
  \item Alicia Schep (\href{https://orcid.org/0000-0002-3915-0618}{ORCID}) [contributor]
  \item Sean Hughes (\href{https://orcid.org/0000-0002-9409-9405}{ORCID}) [contributor]
  \item Trung Kien Dang (\href{https://orcid.org/0000-0001-7562-6495}{ORCID}) [contributor]
  \item Saulius Lukauskas [contributor]
  \item Jean-Olivier Irisson (\href{https://orcid.org/0000-0003-4920-3880}{ORCID}) [contributor]
  \item Zhian N Kamvar (\href{https://orcid.org/0000-0003-1458-7108}{ORCID}) [contributor]
  \item Ryan Thompson (\href{https://orcid.org/0000-0002-0450-8181}{ORCID}) [contributor]
  \item Christophe Dervieux (\href{https://orcid.org/0000-0003-4474-2498}{ORCID}) [contributor]
  \item Hiroaki Yutani [contributor]
  \item Pierre Gramme [contributor]
  \item Amir Masoud Abdol [contributor]
  \item Malcolm Barrett (\href{https://orcid.org/0000-0003-0299-5825}{ORCID}) [contributor]
  \item Robrecht Cannoodt (\href{https://orcid.org/0000-0003-3641-729X}{ORCID}) [contributor]
  \item Michał Krassowski (\href{https://orcid.org/0000-0002-9638-7785}{ORCID}) [contributor]
  \item Michael Chirico (\href{https://orcid.org/0000-0003-0787-087X}{ORCID}) [contributor]
  \item Pedro Aphalo (\href{https://orcid.org/0000-0003-3385-972X}{ORCID}) [contributor]
  \item Francis Barton [contributor]
  \item Yann Cohen (\href{https://orcid.org/0009-0009-0509-3609}{ORCID}) [contributor]
  \item cojiso [contributor]
}

}
\keyword{internal}
