% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{intersect_line_rectangle}
\alias{intersect_line_rectangle}
\title{Find the intersections between a line and a rectangle.}
\usage{
intersect_line_rectangle(p1, p2, b)
}
\arguments{
\item{p1}{A point like \code{c(x, y)}}

\item{p2}{A point like \code{c(x, y)}}

\item{b}{A rectangle like \code{c(x1, y1, x2, y2)}}
}
\description{
Find the intersections between a line and a rectangle.
}

