% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-label-repel.R
\name{makeContent.labelrepelgrob}
\alias{makeContent.labelrepelgrob}
\title{grid::makeContent function for labelRepelGrob.}
\usage{
makeContent.labelrepelgrob(x)
}
\description{
grid::makeContent function for labelRepelGrob.
}

