% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-label-repel.R
\name{makeContent.labelrepeltree}
\alias{makeContent.labelrepeltree}
\title{grid::makeContent function for the grobTree of textRepelGrob objects}
\usage{
\method{makeContent}{labelrepeltree}(x)
}
\arguments{
\item{x}{A grid grobTree.}
}
\description{
grid::makeContent function for the grobTree of textRepelGrob objects
}

