% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-text-repel.R
\name{makeContent.textrepelgrob}
\alias{makeContent.textrepelgrob}
\title{grid::makeContent function for textRepelGrob.}
\usage{
makeContent.textrepelgrob(x)
}
\description{
grid::makeContent function for textRepelGrob.
}

