% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-text-repel.R
\name{makeContent.textrepeltree}
\alias{makeContent.textrepeltree}
\title{grid::makeContent function for the grobTree of textRepelGrob objects}
\usage{
\method{makeContent}{textrepeltree}(x)
}
\arguments{
\item{x}{A grid grobTree.}
}
\description{
grid::makeContent function for the grobTree of textRepelGrob objects
}

