% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{overlaps}
\alias{overlaps}
\title{Test if a box overlaps another box.}
\usage{
overlaps(a, b)
}
\arguments{
\item{a}{A numeric vector representing a box like \code{c(x1, y1, x2, y2)}}

\item{b}{A numeric vector representing a box like \code{c(x1, y1, x2, y2)}}
}
\description{
Test if a box overlaps another box.
}

