% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{overlaps_points}
\alias{overlaps_points}
\title{Test if a box overlaps any point in a list of points.}
\usage{
overlaps_points(b, ps)
}
\arguments{
\item{b}{A numeric vector representing a box like \code{c(x1, y1, x2, y2)}}

\item{ps}{A list of numeric vectors representing points like
\code{list(c(x, y), c(x, y), ...)}}
}
\description{
Test if a box overlaps any point in a list of points.
}

