% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{point_within_box}
\alias{point_within_box}
\title{Test if a point is within the boundaries of a box.}
\usage{
point_within_box(p, b)
}
\description{
Test if a point is within the boundaries of a box.
}

