% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-nudge2.R
\name{position_nudge2}
\alias{position_nudge2}
\title{Nudge points a fixed distance}
\usage{
position_nudge2(x = 0, y = 0)
}
\arguments{
\item{x, y}{Amount of vertical and horizontal distance to move.}
}
\description{
`position_nudge` is generally useful for adjusting the position of
items on discrete scales by a small amount. Nudging is built in to
[geom_text()] because it's so useful for moving labels a small
distance from what they're labelling.
}
\examples{
df <- data.frame(
  x = c(1,3,2,5),
  y = c("a","c","d","c")
)

ggplot(df, aes(x, y)) +
  geom_point() +
  geom_text(aes(label = y))

ggplot(df, aes(x, y)) +
  geom_point() +
  geom_text(aes(label = y), position = position_nudge(y = -0.1))

# Or, in brief
ggplot(df, aes(x, y)) +
  geom_point() +
  geom_text(aes(label = y), nudge_y = -0.1)
}
\concept{position adjustments}
