% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-nudge-repel.R
\name{position_nudge_repel}
\alias{position_nudge_repel}
\title{Nudge labels a fixed distance from points}
\usage{
position_nudge_repel(x = 0, y = 0)
}
\arguments{
\item{x, y}{Amount of vertical and horizontal distance to move.}
}
\description{
`position_nudge_repel` is generally useful for adjusting the starting
position of labels or text to be repelled while preserving the original
position as the start of the segments. Nudging is built in to
[geom_text_repel()] because it is useful for consistently moving labels some
distance from what they're labelling.
}
\examples{
df <- data.frame(
  x = c(1,3,2,5),
  y = c("a","c","d","c")
)

ggplot(df, aes(x, y)) +
  geom_point() +
  geom_text_repel(aes(label = y))

ggplot(df, aes(x, y)) +
  geom_point() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_nudge_repel(x = 0.1, y = 0.15))

# Or, alternatively
ggplot(df, aes(x, y)) +
  geom_point() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  nudge_x = 0.1,
                  nudge_y = 0.15)
}
\concept{position adjustments}
