% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-nudge-repel-t.R
\name{position_nudge_repel_t}
\alias{position_nudge_repel_t}
\title{Nudge labels a variable distance from points}
\usage{
position_nudge_repel_t(
  x = 0,
  y = 0,
  xy_relative = c(0.03, 0.03),
  abline = NULL,
  method = NULL,
  formula = y ~ x,
  direction = NULL,
  line_nudge = 1
)
}
\arguments{
\item{x, y}{Amount of vertical and horizontal distance to move. A numeric
vector of length 1, or of the same length as rows there are in `data`.}

\item{xy_relative}{Nudge relative to _x_ and _y_ data expanse, ignored
unless `x` and `y` are both `NA`s.}

\item{abline}{a vector of length two giving the intercept and slope.}

\item{method}{One of `"spline"`, `"lm"` or `"auto"`.}

\item{formula}{A model formula for [lm()] when `method = "lm"`. Ignored
otherwise.}

\item{direction}{One of "none", or "split".}

\item{line_nudge}{A positive multiplier >= 1, increasing nudging
away from the curve or line compared to nudging from points.}
}
\description{
`position_nudge_repel_t` is generally useful for adjusting the starting
position of labels or text to be repelled while preserving the original
position as the start of the segments. The difference compared to
[position_nudge_repel()] is that the nudging is away from from a line or
curve fitted to the data points or supplied as coefficients. While
[position_nudge_repel()] is most useful for "round-shaped", vertically- or
horizontally elongated clouds of points, [position_nudge_repel_t()] is most
suitable when observations have a linear or curvilinear relationship
between _x_ and _y_ values.
}
\details{
The default ammount of nudging is 3% of the spread of the data along
  _x_ and _y_ axes, which in most cases is good. In most cases it is best to
  apply nudging along a direction perpendicular to the line or curve, if this
  is the aim, passing an argument to only one of `x`, `y` or `xy_relative`
  will be enough. When `direction = "split"` nudging is away from an implicit
  line or curve on either side with positive nudging. The line of curve can
  be smooth spline or linear regression fitted on-the-fly to the data points,
  or a straight line defined by its coefficients passed to `abline`. The
  fitting is well defined only if the observations fall roughly on a curve or
  straight line that is monotonic in `y`. By means of `line_nudge` one can
  increment nudging away from the line or curve compared to away from the
  points, which is useful for example to keep labels outside of a confidence
  band. Direction defaults to `"split"` when `line_nudge > 1`, and otherwise
  to `"none"`.

  This position function is implemented as a panel function and supports
  grouping: the spline or linear model are fitted per group wihin each
  plot panel.
}
\note{
For `method = "lm"` only model formulas corresponding to polynomials
  with no missing terms are supported. If using [poly()], `raw = TRUE` is
  required.

  In practice, `x` and `y` should have the same sign for nudging to work
  correctly.

  This position is most useful when labeling points conforming a cloud along
  an arbitrary curve or line.
}
\examples{

set.seed(16532)
df <- data.frame(
  x = -10:10,
  y = (-10:10)^2,
  yy = (-10:10)^2 + rnorm(21, 0, 4),
  yyy = (-10:10) + rnorm(21, 0, 4),
  l = letters[1:21]
)

# Point on a line or curve

ggplot(df, aes(x, y)) +
  geom_line(linetype = "dotted") +
  geom_point() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 0.3, y = 4)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, y)) +
  geom_line(linetype = "dotted") +
  geom_point() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = -0.5, y = -6)) +
  scale_y_continuous(expand = expansion(mult = 0.35))

ggplot(df, aes(x, y)) +
  geom_line(linetype = "dotted") +
  geom_point() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 0.5, y = 6)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, -y)) +
  geom_line(linetype = "dotted") +
  geom_point() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = -0.5, y = -7)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, -y)) +
  geom_line(linetype = "dotted") +
  geom_point() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 0.5, y = 7)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, y - 40)) +
  geom_line(linetype = "dotted") +
  geom_point() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 0.5, y = 7)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(y, sqrt(y))) +
  geom_line(linetype = "dotted") +
  geom_point() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 7, y = 1)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(y, sqrt(y))) +
  geom_line(linetype = "dotted") +
  geom_point() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = -5, y = -0.5)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, x * 2 + 5)) +
  geom_line(linetype = "dotted") +
  geom_point() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 0.3, y = 3,
                                                    abline = c(5, 2))) +
  scale_y_continuous(expand = expansion(mult = 0.15))

# Points scattered near a curve or line, we use 'direction = "split"'
ggplot(df, aes(x)) +
  geom_line(aes(y = y), linetype = "dotted") +
  geom_point(aes(y = yy)) +
  geom_text_repel(aes(y = yy, label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 0.5, y = 6,
                                                    direction = "split")) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(y, yy)) +
  stat_smooth(method = "lm") +
  geom_point() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 2, y = 8,
                                                    direction = "split")) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, yyy)) +
  stat_smooth(method = "lm") +
  geom_point() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  xlim = c(-12, 12),
                  position = position_nudge_repel_t(x = 0.5, y = 0.5,
                                                    line_nudge = 10,
                                                    direction = "split")) +
  scale_y_continuous(expand = expansion(mult = 0.15)) +
  scale_x_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(y, yy)) +
  stat_smooth(method = "lm") +
  geom_point() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 1, y = 4,
                                                    method = "linear",
                                                    direction = "split")) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(y, yy)) +
  stat_smooth(method = "lm") +
  geom_point() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 1, y = 4,
                                                    abline = c(0, 1),
                                                    direction = "split")) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(y, -yy)) +
  stat_smooth(method = "lm") +
  geom_point() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 1, y = 4,
                                                    method = "linear",
                                                    direction = "split")) +
  scale_y_continuous(expand = expansion(mult = 0.15))

# can also be used with "normal" geoms
ggplot(df, aes(x, y)) +
  geom_line(linetype = "dotted") +
  geom_point() +
  geom_text(aes(label = l),
            position = position_nudge_repel_t(x = 0.3, y = 4)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

}
\seealso{
[ggplot::position_nudge()], [ggrepel::position_nudge_repel()].

Other position adjustments: 
\code{\link{position_nudge_repel}()}
}
\concept{position adjustments}
