% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-r-nudge-repel.R
\name{position_r_nudge_repel}
\alias{position_r_nudge_repel}
\title{Nudge labels a fixed distance from points}
\usage{
position_r_nudge_repel(x = 0, y = 0, x_center = mean, y_center = mean)
}
\arguments{
\item{x, y}{Amount of vertical and horizontal distance to move. A numeric
vector of length 1, or of the same length as rows there are in `data`,}

\item{x_center, y_center}{The coordinates of the virtual origin out from which
nudging radiates, a numeric vector of length 1 or of the same length as
rows there are in `data`, or a function returning either of these computed
from the variable in data.}
}
\description{
`position_r_nudge_repel` is generally useful for adjusting the starting
position of labels or text to be repelled while preserving the original
position as the start of the segments. The difference compared to
[position_nudge_repel()] is that the nudging is radial and away from a focus
point, which by default is the centroid of the data as computed using
[base::mean()]. A different function or a numeric vector can be used to
override this default.
}
\examples{
df <- data.frame(
  x = c(1,3,2,5),
  z = c(4,8,5,6),
  y = c("a","c","d","c")
)

ggplot(df, aes(x, y)) +
  geom_point() +
  geom_text_repel(aes(label = y))

ggplot(df, aes(x, y)) +
  geom_point() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_r_nudge_repel(x = 0.3, y = 0.3)) +
  scale_x_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, z)) +
  geom_point() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_r_nudge_repel(x = 0.3,
                                                    y = 0.6,
                                                    x_center = min,
                                                    y_center = min)) +
  scale_x_continuous(expand = expansion(mult = 0.15)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, z)) +
  geom_point() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_r_nudge_repel(x = 0.5,
                                                    y = 0.5,
                                                    x_center = 2.5,
                                                    y_center = 6)) +
  scale_x_continuous(expand = expansion(mult = 0.15)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, z)) +
  geom_point() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_r_nudge_repel(x = 0.3, y = 0.6,
                                                    x_center = 3,
                                                    y_center = 6)) +
  scale_x_continuous(expand = expansion(mult = 0.15)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

}
\seealso{
Other position adjustments: 
\code{\link{position_nudge_repel}()},
\code{\link{position_s_nudge_repel}()}
}
\concept{position adjustments}
