% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-nudge-split.R
\name{position_s_nudge_repel}
\alias{position_s_nudge_repel}
\title{Nudge labels a fixed distance from points}
\usage{
position_s_nudge_repel(x = 0, y = 0, x_split = mean, y_split = mean)
}
\arguments{
\item{x, y}{Amount of vertical and horizontal distance to move. A numeric
vector of length 1, or of the same length as rows there are in `data`,}

\item{x_split, y_split}{Divides at which the sign of nudge values changes, a
numeric vector of length 1 or of the same length as rows there are in
`data`, or a function returning either of these computed from the variable
in data.}
}
\description{
`position_s_nudge_repel` is generally useful for adjusting the starting
position of labels or text to be repelled while preserving the original
position as the start of the segments. The difference compared to
[position_nudge_repel()] is that the sign of the applied nudge depends
on which side of a divide a given point is. By default the split as at
the centroid of the data as computed using [base::mean()]. A different
function or a numeric vector can be used to override this default.
}
\examples{
df <- data.frame(
  x = c(1,3,2,5),
  y = c("a","c","d","c")
)

ggplot(df, aes(x, y)) +
  geom_point() +
  geom_text_repel(aes(label = y))

ggplot(df, aes(x, y)) +
  geom_point() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_s_nudge_repel(x = 0.1, y = 0.15))

ggplot(df, aes(x, y)) +
  geom_point() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  force = 0,
                  force_pull = 0,
                  position = position_s_nudge_repel(x = 0.1, y = 0.15))

ggplot(df, aes(x, y)) +
  geom_point() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_s_nudge_repel(x = 0.1,
                                                    y = 0.15,
                                                    x_split = 1.5))

}
\seealso{
Other position adjustments: 
\code{\link{position_nudge_repel}()}
}
\concept{position adjustments}
