% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-t-nudge-repel.R
\name{position_t_nudge_repel}
\alias{position_t_nudge_repel}
\title{Transverse nudge labels a fixed distance from points}
\usage{
position_t_nudge_repel(x = 0, y = 0, ...)
}
\arguments{
\item{x, y}{Amount of vertical and horizontal distance to move. A numeric
vector of length 1, or of the same length as rows there are in `data`,}

\item{...}{Additional arguments passed to `smooth.spline()`.}
}
\description{
`position_t_nudge_repel` is generally useful for adjusting the starting
position of labels or text to be repelled while preserving the original
position as the start of the segments. The difference compared to
[position_nudge_repel()] is that the nudging is away from from a smooth
line along the data points.
}
\note{
Nudging direction is away from a smooth spline fitted to the data
  points. This works only if the observations fall roughly on a curve that is
  montonic in `y`.
}
\examples{
set.seed(16532)
df <- data.frame(
  x = -10:10,
  z = (-10:10)^2,
  y = letters[1:21]
)

ggplot(df, aes(x, z)) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_t_nudge_repel(x = 1, y = 12)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, z)) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_t_nudge_repel(x = -0.5, y = -6)) +
  scale_y_continuous(expand = expansion(mult = 0.35))

ggplot(df, aes(x, z)) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_t_nudge_repel(x = 0.5, y = 7,
                                                    x_center = mean,
                                                    y_center = max)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, -z)) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_t_nudge_repel(x = -0.5, y = -7,
                                                    x_center = mean,
                                                    y_center = max)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, -z)) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_t_nudge_repel(x = 0.5, y = 7,
                                                    x_center = mean,
                                                    y_center = max)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, z - 40)) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_t_nudge_repel(x = 0.5, y = 7,
                                                    x_center = mean,
                                                    y_center = max)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(z, sqrt(z))) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_t_nudge_repel(x = 7, y = 1,
                                                    x_center = mean,
                                                    y_center = max)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(z, sqrt(z))) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_t_nudge_repel(x = -5, y = -0.5,
                                                    x_center = mean,
                                                    y_center = max)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

}
\seealso{
Other position adjustments: 
\code{\link{position_nudge_repel}()},
\code{\link{position_r_nudge_repel}()},
\code{\link{position_s_nudge_repel}()}
}
\concept{position adjustments}
