% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repel_boxes}
\alias{repel_boxes}
\title{Adjust the layout of a list of potentially overlapping boxes.}
\usage{
repel_boxes(boxes, xlim, ylim, force = 1e-06, maxiter = 10000L)
}
\description{
Adjust the layout of a list of potentially overlapping boxes.
}

