% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repel_boxes.R
\name{repel_boxes_R}
\alias{repel_boxes_R}
\title{Adjust the layout of a list of potentially overlapping boxes.}
\usage{
repel_boxes_R(boxes, xlim = c(-Inf, Inf), ylim = c(-Inf, Inf),
  force = 1e-06, maxiter = 10000)
}
\description{
Adjust the layout of a list of potentially overlapping boxes.
}

