% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repel_force}
\alias{repel_force}
\title{Compute the repulsion force upon point \code{a} from point \code{b}.}
\usage{
repel_force(a, b, force = 1e-06)
}
\arguments{
\item{a}{A point like \code{c(x, y)}}

\item{b}{A point like \code{c(x, y)}}

\item{force}{Magnitude of the force (defaults to \code{1e-6})}
}
\description{
The force decays with the squared distance between the points, similar
to the force of repulsion between magnets.
}

