% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spring_force}
\alias{spring_force}
\title{Compute the spring force upon point \code{a} from point \code{b}.}
\usage{
spring_force(a, b, force = 1e-06)
}
\arguments{
\item{a}{A point like \code{c(x, y)}}

\item{b}{A point like \code{c(x, y)}}

\item{force}{Magnitude of the force (defaults to \code{1e-6})}
}
\description{
The force increases with the distance between the points, similar
to Hooke's law for springs.
}

