% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wordcloud_layout.R
\name{wordcloud_layout}
\alias{wordcloud_layout}
\title{Adjust the layout of a list of potentially overlapping boxes.}
\usage{
wordcloud_layout(boxes, xlim = c(-Inf, Inf), ylim = c(-Inf, Inf),
  tstep = 0.01, rstep = 0.01, maxiter = 10000)
}
\description{
Adjust the layout of a list of potentially overlapping boxes.
}

