% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{box_within_bounds}
\alias{box_within_bounds}
\title{Test if a box is within xlim and ylim boundaries.}
\usage{
box_within_bounds(b, xlim, ylim)
}
\description{
Test if a box is within xlim and ylim boundaries.
}

