% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-image-repel.R, R/geom-label-repel.R, R/geom-point-repel.R, R/geom-text-repel.R
\docType{data}
\name{GeomImageRepel}
\alias{GeomImageRepel}
\alias{GeomLabelRepel}
\alias{GeomPointRepel}
\alias{GeomTextRepel}
\title{GeomImageRepel}
\description{
GeomImageRepel

GeomLabelRepel

GeomTextRepel

GeomTextRepel
}
\keyword{datasets}
