% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-label-repel.R, R/geom-text-repel.R,
%   R/geom-text-wordcloud.R
\docType{data}
\name{GeomLabelRepel}
\alias{GeomLabelRepel}
\alias{GeomTextRepel}
\alias{GeomTextWordcloud}
\title{GeomLabelRepel}
\description{
GeomLabelRepel

GeomTextRepel

GeomTextWordcloud
}
\keyword{datasets}
