% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-label-repel.R, R/geom-text-repel.R
\docType{data}
\name{GeomLabelRepel}
\alias{GeomLabelRepel}
\alias{GeomTextRepel}
\title{GeomLabelRepel}
\description{
GeomLabelRepel

GeomTextRepel
}
\seealso{
\link[ggplot2]{GeomLabel} from the ggplot2 package.

\link[ggplot2]{GeomText} from the ggplot2 package.
}
\keyword{datasets}
