% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrepel-package.R
\docType{package}
\name{ggrepel}
\alias{ggrepel}
\alias{ggrepel-package}
\title{ggrepel}
\description{
This package contains extra geoms for \pkg{ggplot2}.
}
\details{
In addition to the help pages, see the README page on
\href{https://github.com/slowkow/ggrepel}{github} for examples.
}

