% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggrepel-package.R
\docType{package}
\name{ggrepel}
\alias{ggrepel}
\alias{ggrepel-package}
\title{ggrepel}
\description{
This package contains extra geoms for \pkg{ggplot2}.
}
\details{
Please see the help pages listed below:

\itemize{
  \item \code{\link{geom_text_repel}}
  \item \code{\link{geom_label_repel}}
}

Also see the vignette for more usage examples:

\code{browseVignettes("ggrepel")}

Please report issues and suggest improvements at Github:

\url{https://github.com/slowkow/ggrepel}
}
