% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{point_within_box}
\alias{point_within_box}
\title{Test if a point is within the boundaries of a box.}
\usage{
point_within_box(p, b)
}
\arguments{
\item{p}{A point like \code{c(x, y)}}

\item{b}{A numeric vector representing a box like \code{c(x1, y1, x2, y2)}}
}
\description{
Test if a point is within the boundaries of a box.
}

