% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-nudge2.R
\name{position_nudge_repel}
\alias{position_nudge_repel}
\title{Nudge points a fixed distance}
\usage{
position_nudge_repel(x = 0, y = 0)
}
\arguments{
\item{x, y}{Amount of vertical and horizontal distance to move.}
}
\description{
`position_nudge_repel` is generally useful for adjusting the starting
position of labels or text to be repelled while preserving the original
position as the start of the segments. Nudging is built in to
[geom_text_repel()] because it is useful for consistently moving labels some
distance from what they're labelling.
}
\concept{position adjustments}
