% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-nudge-repel-t.R
\name{position_nudge_repel_t}
\alias{position_nudge_repel_t}
\title{Transverse nudge labels a fixed distance from points}
\usage{
position_nudge_repel_t(x = 0, y = 0, method = "spline", direction = "none")
}
\arguments{
\item{x, y}{Amount of vertical and horizontal distance to move. A numeric
vector of length 1, or of the same length as rows there are in `data`,}

\item{method}{One of "spline" or "linear".}

\item{direction}{One of "none", or "split".}
}
\description{
`position_nudge_repel_t` is generally useful for adjusting the starting
position of labels or text to be repelled while preserving the original
position as the start of the segments. The difference compared to
[position_nudge_repel()] is that the nudging is away from from a smooth
line along the data points.
}
\note{
Nudging direction is away from a smooth spline or linear regression
  fitted to the data points. This works only if the observations fall roughly
  on a curve or straight line that is montonic in `y`.
}
\examples{

set.seed(16532)
df <- data.frame(
  x = -10:10,
  y = (-10:10)^2,
  yy = (-10:10)^2 + rnorm(21, 0, 4),
  l = letters[1:21]
)

# Point on a line or curve

ggplot(df, aes(x, y)) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 1, y = 12)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, y)) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = -0.5, y = -6)) +
  scale_y_continuous(expand = expansion(mult = 0.35))

ggplot(df, aes(x, y)) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 0.5, y = 7)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, -y)) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = -0.5, y = -7)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, -y)) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 0.5, y = 7)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, y - 40)) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 0.5, y = 7)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(y, sqrt(y))) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 7, y = 1)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(y, sqrt(y))) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = -5, y = -0.5)) +
  scale_y_continuous(expand = expansion(mult = 0.15))

# Points scattered near a curve or line, we use 'direction = "split"'
ggplot(df, aes(x)) +
  geom_point(aes(y = yy)) +
  geom_line(aes(y = y)) +
  geom_text_repel(aes(y = yy, label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 1, y = 12,
                                                    direction = "split")) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(y, yy)) +
  geom_point() +
  stat_smooth(method = "lm") +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 3, y = 3,
                                                    direction = "split")) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(y, yy)) +
  geom_point() +
  stat_smooth(method = "lm") +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 3, y = 3,
                                                    method = "linear",
                                                    direction = "split")) +
  scale_y_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(y, -yy)) +
  geom_point() +
  stat_smooth(method = "lm") +
  geom_text_repel(aes(label = l),
                  min.segment.length = 0,
                  position = position_nudge_repel_t(x = 3, y = 3,
                                                    method = "linear",
                                                    direction = "split")) +
  scale_y_continuous(expand = expansion(mult = 0.15))

}
\seealso{
Other position adjustments: 
\code{\link{position_nudge_repel}()}
}
\concept{position adjustments}
