% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/position-r-nudge-repel.R
\name{position_r_nudge_repel}
\alias{position_r_nudge_repel}
\title{Radially nudge labels a fixed distance from points}
\usage{
position_r_nudge_repel(x = 0, y = 0, x_center = mean, y_center = mean)
}
\arguments{
\item{x, y}{Amount of vertical and horizontal distance to move. A numeric
vector of length 1, or of the same length as rows there are in `data`,}

\item{x_center, y_center}{The coordinates of the virtual origin out from which
nudging radiates, a numeric vector of length 1 or of the same length as
rows there are in `data`, or a function returning either of these computed
from the variable in data.}
}
\description{
`position_r_nudge_repel` is generally useful for adjusting the starting
position of labels or text to be repelled while preserving the original
position as the start of the segments. The difference compared to
[position_nudge_repel()] is that the nudging is radial and away from a focus
point, which by default is the centroid of the data as computed using
[base::mean()]. A different function or a numeric vector can be used to
override this default.
}
\note{
Positive values as arguments to `x` and `y` nudge
  radialy outwards from the centre, while negative values nudge inwards
  towards the center. The segment length equal to `y` for vertical
  segments and equal to `x` for horizontal segments. The length of those
  at intermediate angles varies smoothly between these two maximum and
  minimum lengths. Nudging away from a vertical or horizontal line
  is implemented in `position_s_nudge_repel()`.
}
\examples{
df <- data.frame(
  x = -10:10,
  z = (-10:10)^2,
  y = letters[1:21]
)

ggplot(df, aes(x, z)) +
  geom_point() +
  geom_text_repel(aes(label = y))

ggplot(df, aes(x, z)) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_r_nudge_repel(x = -1, y = -6,
                                                    x_center = mean,
                                                    y_center = max)) +
  scale_x_continuous(expand = expansion(mult = 0.15))

ggplot(df, aes(x, z)) +
  geom_point() +
  geom_line() +
  geom_text_repel(aes(label = y),
                  min.segment.length = 0,
                  position = position_r_nudge_repel(x = 1, y = 6,
                                                    x_center = mean,
                                                    y_center = max)) +
  scale_y_continuous(expand = expansion(mult = 0.15)) +
  scale_x_continuous(expand = expansion(mult = 0.15))

}
\seealso{
Other position adjustments: 
\code{\link{position_nudge_repel}()},
\code{\link{position_s_nudge_repel}()},
\code{\link{position_t_nudge_repel}()}
}
\concept{position adjustments}
