% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{repel_boxes}
\alias{repel_boxes}
\title{Adjust the layout of a list of potentially overlapping boxes.}
\usage{
repel_boxes(data_points, boxes, xlim, ylim, force = 1e-06, maxiter = 2000L)
}
\arguments{
\item{data_points}{A numeric matrix with rows representing points like
\code{rbind(c(x, y), c(x, y), ...)}}

\item{boxes}{A numeric matrix with rows representing boxes like
\code{rbind(c(x1, y1, x2, y2), c(x1, y1, x2, y2), ...)}}

\item{xlim}{A numeric vector representing the limits on the x axis like
\code{c(xmin, xmax)}}

\item{ylim}{A numeric vector representing the limits on the y axis like
\code{c(ymin, ymax)}}

\item{force}{Magnitude of the force (defaults to \code{1e-6})}

\item{maxiter}{Maximum number of iterations to try to resolve overlaps
(defaults to 2000)}
}
\description{
Adjust the layout of a list of potentially overlapping boxes.
}

